add screens for auto and custom collections
This adds configuration for auto collections so they do not use the
default theme, a theme for the "group unthemed custom collections"
configuration, as well as a default theme so that when custom
collections are ungrouped, they still look nice.

The auto collections use icons from Font Awesome along with text drawn
into the SVG using
https://danmarshall.github.io/google-font-to-svg-path/ (since I could
not get <text> elements to work in the SVGs...)

All of these screens use the RetroPie background as I could not think of
any better background. A simple radial gradient background could work
nicely though...